+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Add `attributedTo` property to `Emoji` objects.
+- Support integrity proofs with injected context.
 
 ### Changed